requirements.txt
ipykernel
langchain
langgraph
langchain-groq
langchain-community
sentence-transformers
chromadb
python-dotenv
typing-extensions
matplotlib
langchain-huggingface 
transformers
huggingface-hub
agno
duckduckgo-search
langchain-chroma
langchain-ollama
openai
streamlit
# ConversationalRetrievalQA
streamlit-chat
youtube_transcript_api
bs4
selenium
yfinance
schedule